refactor: extract companies and stream routes to route files
- Extract POST /companies, POST /link-company, GET /companies/:id/invite-code to routes/companies.js
- Extract GET /stream-token, GET /stream-unread, OPTIONS+POST /sync-stream-user, POST /sync-stream-users to routes/stream.js
- Replace streamServer local var with streamServerClient from ../streamServerClient in stream routes
- Mount both routers via app.use("/api", ...) in server.js
